refactor: move observer to pattterns folder
