+++ lib/patterns/observer.js
