+++ lib/services/post.js
@@ -1,4 +1,4 @@
-import { Observer } from "@lib/services/observer";
+import { Observer } from "@lib/patterns/observer";
 import UserService from "@lib/services/user";
 import { mockPosts } from "@/__mocks__/posts";
 
@@ -10,12 +10,13 @@ class PostService extends Observer {
   constructor() {
     super();
 
-    // 싱글톤 패턴 구현
-    if (PostService.instance) {
-      return PostService.instance;
-    }
+    if (PostService.instance) return PostService.instance;
     PostService.instance = this;
 
+    this.init();
+  }
+
+  init() {
     this.posts = [];
     this.loadPosts();
   }
